+++ src/barfi/v0/block_builder.py
@@ -1,6 +1,6 @@
 import types
 from typing import Callable, Any
-from barfi.deprecated.option_builder import build_option
+from barfi.v0.option_builder import build_option
 
 
 class Block(object):
