+++ src/barfi/v0/static/index.html
