+++ ui-v0-deprecated/src/components/BlockEditor.vue
